move the epic to their more appropriate file
