+++ web/client/epics/__tests__/identify-test.js
@@ -10,10 +10,11 @@ const expect = require('expect');
 
 const { ZOOM_TO_POINT, clickOnMap } = require('../../actions/map');
 const { FEATURE_INFO_CLICK, UPDATE_CENTER_TO_MARKER, PURGE_MAPINFO_RESULTS, NEW_MAPINFO_REQUEST, LOAD_FEATURE_INFO, NO_QUERYABLE_LAYERS, ERROR_FEATURE_INFO, EXCEPTIONS_FEATURE_INFO, SHOW_MAPINFO_MARKER, HIDE_MAPINFO_MARKER, GET_VECTOR_INFO, loadFeatureInfo, featureInfoClick, closeIdentify, toggleHighlightFeature } = require('../../actions/mapInfo');
-const { getFeatureInfoOnFeatureInfoClick, zoomToVisibleAreaEpic, onMapClick, closeFeatureAndAnnotationEditing, handleMapInfoMarker, featureInfoClickOnHighligh } = require('../identify');
+const { getFeatureInfoOnFeatureInfoClick, zoomToVisibleAreaEpic, onMapClick, closeFeatureAndAnnotationEditing, handleMapInfoMarker, featureInfoClickOnHighligh, closeFeatureInfoOnCatalogOpenEpic } = require('../identify');
 const { CLOSE_ANNOTATIONS } = require('../../actions/annotations');
 const { testEpic, TEST_TIMEOUT, addTimeoutEpic } = require('./epicTestUtils');
 const { registerHook } = require('../../utils/MapUtils');
+const { setControlProperties } = require('../../actions/controls');
 
 const TEST_MAP_STATE = {
     present: {
@@ -493,5 +494,37 @@ describe('identify Epics', () => {
                 }
             }
         });
+    });
+
+    it('enable metadataexplorer control is enabled', (done) => {
+        const state = {controls: {}};
+        const NUMBER_OF_ACTIONS = 2;
+        const callback = actions => {
+            expect(actions.length).toBe(NUMBER_OF_ACTIONS);
+            expect(actions[0].type).toEqual(PURGE_MAPINFO_RESULTS);
+            expect(actions[1].type).toEqual(HIDE_MAPINFO_MARKER);
+            done();
+        };
+        testEpic(
+            closeFeatureInfoOnCatalogOpenEpic,
+            NUMBER_OF_ACTIONS,
+            setControlProperties('metadataexplorer', "enabled", true),
+            callback,
+            state);
+    });
+
+    it('disable metadataexplorer should not affect mapinfo', (done) => {
+        const state = {controls: {}};
+        const NUMBER_OF_ACTIONS = 1;
+        const callback = actions => {
+            expect(actions.length).toBe(NUMBER_OF_ACTIONS);
+            done();
+        };
+        testEpic(
+            addTimeoutEpic(closeFeatureInfoOnCatalogOpenEpic),
+            NUMBER_OF_ACTIONS,
+            setControlProperties('metadataexplorer', "enabled", false),
+            callback,
+            state);
     });
 });
