+++ web/client/epics/controls.js
@@ -7,8 +7,6 @@
  */
 
 const Rx = require('rxjs');
-const { SET_CONTROL_PROPERTIES } = require('../actions/controls');
-const { purgeMapInfoResults, hideMapinfoMarker } = require('../actions/mapInfo');
 
 module.exports = {
     onEpic: (action$, store) =>
@@ -18,13 +16,5 @@ module.exports = {
                 return Rx.Observable.of(action.action);
             }
             return Rx.Observable.of(action.elseAction.call());
-        }),
-
-    setControlPropertiesEpic: (action$) =>
-        action$
-            .ofType(SET_CONTROL_PROPERTIES)
-            .filter((action) => action.control === "metadataexplorer" && action.properties && action.properties.enabled)
-            .switchMap(() => {
-                return Rx.Observable.of(purgeMapInfoResults(), hideMapinfoMarker());
-            })
+        })
 };
